requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Werkzeug==2.3.7
python-dotenv==1.0.0
psycopg2-binary==2.9.7
boto3==1.28.36
Pillow==10.0.0
redis==5.0.0
gunicorn==21.2.0
Flask-Migrate==4.0.7
Flask-WTF==1.2.1
email-validator==2.1.0.post1
